+++ openstack_types/src/block_storage/v3/snapshot/response/create.rs
@@ -90,7 +90,7 @@ pub struct SnapshotResponse {
 
     /// The size of the volume, in gibibytes (GiB).
     #[structable()]
-    pub size: i64,
+    pub size: i32,
 
     /// The status for the snapshot.
     #[structable(serialize)]
